refactor(canvas): split chat page and rail styles
