+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatSessionsRail/index.tsx
@@ -1,14 +1,15 @@
 import { useEffect, useId, useMemo, useRef, useState } from 'react';
+import './index.css';
 import {
   ChevronRightIcon,
   PlusIcon,
   SpinnerIcon,
-} from '../icons';
-import { Button, TextField } from '../ui';
-import { useI18n } from '../../i18n';
+} from '../../icons';
+import { Button, TextField } from '../../ui';
+import { useI18n } from '../../../i18n';
 import { ChatSessionRailItem } from './ChatSessionRailItem';
-import type { WorkspaceOption } from './types';
-import type { ConversationCompletionStatus } from '../../agent-chat/runtime/conversationCompletionStore';
+import type { WorkspaceOption } from '../types';
+import type { ConversationCompletionStatus } from '../../../agent-chat/runtime/conversationCompletionStore';
 
 const SESSION_PREVIEW_LIMIT = 10;
 const GLOBAL_CHAT_ID = '__global_chat__';
