+++ apps/canvas-workspace/src/renderer/src/components/chat/__tests__/ChatPageResponsive.test.ts
@@ -1,7 +1,7 @@
 import { readFileSync } from 'node:fs';
 import { describe, expect, it } from 'vitest';
 
-const chatPageCss = readFileSync(new URL('../ChatPage.css', import.meta.url), 'utf8');
+const chatPageCss = readFileSync(new URL('../ChatPageBody/index.css', import.meta.url), 'utf8');
 const rightDockCss = readFileSync(new URL('../../dock/RightDock/index.css', import.meta.url), 'utf8');
 const rendererCss = readFileSync(new URL('../../../styles.css', import.meta.url), 'utf8');
